WIP: Add a threadpool to the sensor daemon
Summary:
This adds a ThreadPoolExecutor to the daemon loop, to allow sensors to run in parallel instead of sequentially. We now make a copy of the workspace on each sensor iteration, since the workspace might get cleaned up while the sensor is in the middle of executing.

Test Plan: BK